+++ src/Attribute/AsCommandArgument.php
@@ -5,7 +5,7 @@ namespace Castor\Attribute;
 abstract class AsCommandArgument
 {
     public function __construct(
-        public readonly string|null $name = null,
+        public readonly ?string $name = null,
     ) {
     }
 }